♻️ Refactor :  커피챗 등록, 수정 로직 단순화
