+++ src/pages/CoffeeUpdatePage.jsx
@@ -1,69 +1,13 @@
-import { useEffect } from 'react';
-
-import { getCoffeeChatDetail, updateCoffeechat } from 'api/api';
-import NewBtn from 'components/common/button/NewBtn';
-import NewDayPicker from 'components/common/date-picker/NewDayPicker';
+import UpdateCoffeeChatForm from 'components/coffeechats/Update/ui/UpdateCoffeeChatForm';
 import Header from 'components/common/header/Header';
-import NewInput from 'components/common/input/NewInput';
+import FormWrapper from 'components/common/wrapper/FormWrapper';
 import { COFFEE_CHILD_ID } from 'constants/headerProps';
-import { useCoffeeForm } from 'hooks/useCoffeeForm';
-import { useNavigate, useParams } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-import { isLoggedInState } from 'recoil/atoms';
-import { theme } from 'styles/themeMuiStyle';
-import { formatDateTime } from 'utils/dateUtils';
+import { useParams } from 'react-router-dom';
 
-import { Box, Container, CssBaseline, Grid, Typography } from '@mui/material';
+import { Box, Container, CssBaseline } from '@mui/material';
 
 function CoffeeUpdatePage() {
   const { id } = useParams();
-  const navigate = useNavigate();
-  const loginState = useRecoilValue(isLoggedInState);
-  const { formValue, setFormValue, helperTexts, isFormValid, updateFormValue } = useCoffeeForm({
-    title: '',
-    content: '',
-    totalRecruitCount: '',
-    meetDate: '',
-    openChatUrl: '',
-  });
-
-  useEffect(() => {
-    (async () => {
-      try {
-        const res = await getCoffeeChatDetail(id);
-        setFormValue({
-          title: res.title || '',
-          content: res.content || '',
-          totalRecruitCount: res.totalRecruitCount || '',
-          meetDate: res.meetDate || '',
-          openChatUrl: res.openChatUrl || '',
-        });
-        if (res.hostId !== loginState.memberId) {
-          alert('본인이 작성하지 않은 커피챗의 상태를 변경할 수 없습니다.');
-          navigate('/coffee');
-        }
-      } catch (error) {
-        if (error.response.status) {
-          navigate('/404');
-        }
-        console.error('Error fetching coffee chat detail:', error);
-      }
-    })();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [id, navigate, setFormValue]);
-
-  const hanldeRegister = async (e) => {
-    e.preventDefault();
-    try {
-      await updateCoffeechat(id, formValue);
-      alert('커피챗이 수정되었습니다.');
-      navigate(`/coffee/${id}`);
-    } catch (error) {
-      const { message } = error.response.data;
-      alert(message);
-      console.error('Error fetching hot skills:', error);
-    }
-  };
 
   return (
     <>
@@ -71,89 +15,9 @@ function CoffeeUpdatePage() {
         <CssBaseline />
         <Header title={COFFEE_CHILD_ID.title} url={COFFEE_CHILD_ID.url + id} />
         <Box flexGrow={1} display="flex" flexDirection="column" justifyContent="center">
-          <Typography fontSize={18} fontWeight={600} paddingTop={1}>
-            수정하실 커피챗 정보를 입력해주세요 ☕️
-          </Typography>
-          <Box
-            component="form"
-            sx={{
-              mt: '30px',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: '13px',
-              width: '100%',
-            }}>
-            <NewInput
-              placeholder="커피챗 제목을 입력해주세요"
-              label="제목"
-              helperText={helperTexts.title}
-              value={formValue.title}
-              onChange={(e) => updateFormValue('title', e.target.value)}
-            />
-            <NewInput
-              placeholder="커피챗 내용을 입력해주세요"
-              label="상세 내용"
-              helperText={helperTexts.content}
-              value={formValue.content}
-              isMultiline={true}
-              onChange={(e) => updateFormValue('content', e.target.value)}
-            />
-            <Box>
-              <Grid container width="100%" display="flex" justifyContent="space-between">
-                <Grid item xs={5.6}>
-                  <NewInput
-                    placeholder="숫자만 입력해주세요"
-                    label="총 모집 인원"
-                    helperText={helperTexts.totalRecruitCount}
-                    type="number"
-                    min={0}
-                    value={
-                      isNaN(formValue.totalRecruitCount) || formValue.totalRecruitCount === 0
-                        ? null
-                        : formValue.totalRecruitCount
-                    }
-                    onChange={(e) => {
-                      updateFormValue('totalRecruitCount', parseInt(e.target.value, 10));
-                    }}
-                  />
-                </Grid>
-
-                <Grid item xs={5.6}>
-                  <NewDayPicker
-                    label="일시"
-                    daytime={true}
-                    value={formValue.meetDate}
-                    onChange={(newValue) => {
-                      updateFormValue('meetDate', formatDateTime(newValue));
-                    }}
-                  />
-                </Grid>
-              </Grid>
-            </Box>
-            <NewInput
-              placeholder="오픈채팅방 링크를 입력해주세요"
-              label="오픈채팅방 링크"
-              helperText={helperTexts.openChatUrl}
-              value={formValue.openChatUrl}
-              onChange={(e) => updateFormValue('openChatUrl', e.target.value)}
-            />
-            <NewBtn
-              title="수정완료"
-              onClick={hanldeRegister}
-              disable={!isFormValid}
-              isActive={!isFormValid}
-              styles={{
-                mt: 3,
-                mb: 3,
-                width: '100%',
-                p: '13px',
-                borderRadius: '999px',
-                background: isFormValid ? theme.palette.primary.main : '#EBEBEB',
-                color: isFormValid ? 'white' : '#BCBCC4',
-                fontSize: '16px',
-              }}
-            />
-          </Box>
+          <FormWrapper title="수정하실 커피챗 정보를 입력해주세요. ☕️" styles={{ mt: '30px' }}>
+            <UpdateCoffeeChatForm />
+          </FormWrapper>
         </Box>
       </Container>
     </>
